ml/requirements.txt
catboost>=1.2
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
psycopg2-binary>=2.9
